requirements.txt
rdkit-pypi
pmapper
numpy
pymol-open-source
